+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 ## [Unreleased]
 
 - **feat**: enhance path handling for moved workspace in problem data
+- **feat**: add javascript language support
 
 ## 0.4.8